Proposed changes to fix the description of DNS
DNS doesn't deal with URLs. It deals with host names. The picture is also inaccurate: A DNS request would be sent to mycluster.us-south.containers.appdomain.cloud, not mycluster.us-south.containers.appdomain.cloud/myapp. Also, DNS does not return port numbers as incorrectly suggested by the picture. It returns an IP address for a hostname. The port number and URL path are NOT part of DNS.